+++ src/webpack.worlds.parts/circles_assets.part.html
@@ -7,6 +7,7 @@
       <a-entity   class="avatar" 
                   position="0 __USER_HEIGHT__ 0"
                   circles-user-networked
+                  circles-user-local
                   circles-avatar-construction>
         <a-entity class="nametag_front avatar_node" text="align:center;" position="0 0.7 0" rotation="0 0 0" scale="3 3 3">
           <a-entity class="deviceicon_front" position="0 -0.06 0" geometry="primitive:plane; width:0.07; height:0.07;" material="shader:flat; transparent:true;"></a-entity>
