+++ src/webpack.worlds.parts/circles_end_scripts.part.html
@@ -116,18 +116,6 @@
         component: 'position',
       },
       // {
-      //   selector: '.user_head',
-      //   component: 'circles-shader',
-      // },
-      // {
-      //   selector: '.user_hair',
-      //   component: 'circles-shader',
-      // },
-      // {
-      //   selector: '.user_body',
-      //   component: 'circles-shader',
-      // },
-      // {
       //   selector: '.user_body',
       //   component: 'rotation',
       // },
